+++ src/app.module.ts
@@ -5,8 +5,8 @@ import { AuthModule } from './auth/auth.module';
 import { UserModule } from './user/user.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { ConfigModule } from '@nestjs/config';
-import { PostService } from './post/post.service';
-import { PostController } from './post/post.controller';
+import { PostService } from './posts/post.service';
+import { PostController } from './posts/post.controller';
 
 @Module({
   imports: [
