+++ src/posts/dto/update-post.dto.ts
@@ -1,4 +1,4 @@
-import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
+import { IsNotEmpty, IsString } from 'class-validator';
 
 export class UpdatePostDto {
   @IsNotEmpty()
