+++ src/posts/interfaces/post.interface.ts
